services/user-auth/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
coverage
pytest-cov
gunicorn
pyjwt
python-multipart
